backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Validation & settings
pydantic==2.10.3
pydantic-settings==2.7.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP client (external APIs)
httpx==0.28.1
requests==2.31.0

# ML / Data
pandas==2.2.3
scikit-learn==1.6.0
xgboost==2.1.3
joblib==1.4.2
numpy==2.2.0

# PDF generation
reportlab==4.2.5

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Code quality
ruff==0.8.4
